backend/ml/p/requirements.txt
pandas
numpy
scikit-learn
pymongo
flask
joblib
python-dotenv
matplotlib
seaborn